Iterative solution using BFS and Deque
Using Deque from collections is more efficient than list and pop(0).
Here inside while loop we change the size every time for a new level.
We run for loop for all the elements on that level till all the children
of that level are added. Size increases with level. Time complexity is
O(n) and Space complexity is O(2n). In other solutions, we add the node
value directly to the res list at the appropriate level. Here we add
the node value to a new list. And then add that list to the output after
all elements at that level have been added.