Recursive solution using max
Similar to max depth, one condition where either of left and right node
is None, needs to be handled. In max depth, we return max of left and
right nodes. In min depth, min of left and right. However, if one of
them is None then return the max of either.