Iterative solution using next function
Whenever we come across a duplicate, we will assign head.next to the
next to next element. We can use 2 variables or single variable for this
approach. While loop runs until head.next is not None. Check for edge
case of head is None. Use a dummy variable to store head and operate on
it, return head at the end because Python stores by reference.