3 pointers solution
Implemented because it is intutive. There are 3 pointers used i, j, k.
i is to attach child (k) to the end. j is to attach child's tail to the
flattened list. This approach has the disadvantage that we will have to
traverse the same sublist twice. So instead of O(n), time complexity
becomes O(2n).