Using recursion and dfs
Here, similar to intutive solution instead of visiting sublist twice, we
only visit once. Therefore, time complexity is O(N). We store curr.next
in a temporary variable and then use it later because the value of curr.
next changes after attaching curr.child's sublist to the curr element.
To attach curr.child as next of curr, we use recursion wherein (prev,
curr) is portrayed by (curr, curr.child). Make sure to remove curr.child
later on. Store the result which is the tail of child sublist. Again run
recursion on (tail, tempNext) and return that result. So basically this
method returns the tail of running recursion.
The main function needs declaration of pseudoHead before head so that it
can be used as prev in helper function. Make sure to remove that pointer
before returning head otherwise it gives 'mylist' is not a valid doubly
linked list.