Own approach - make ring, assign new head and tail
To shift back elements to the front we need to make linked list into a
cycle. Therefore, we connect the current tail to the current head of the
linked list. And then will break the ring by determining the location of
new head and new tail.
3 side cases - 1. list is empty 2. list has one
element 3. k = 0, i.e no shifting needed
One more optimization is making k less than length of linked list.
Therefore, first step is to determine the length of linked list. We do
that by iterating over the linked list and counting the elements. If k%n
turns out 0, we return the list as it is.
We again iterate using the same pointer which is at the end of list till
count reaches n-k to find the new head and tail. There should be k back
elements in front of the list. Therefore there should be n-k front
elements at the end of list. Eg. n = 5, k = 1, count = 0, pointer is at
5, count runs from 1,2,3,4. pointer.next i.e 5 becomes head and None is
assigned to 4.next.
time complexity: O(n)
space complexity: O(1)