Make cycle, find new tail and head
Similar to own approach. Take care of 3 base cases, 1. list is empty,
2. list has one element, 3. k = 0. Iterate over the entire linked list
to get the length. Start from head to iterate over the linked list again
till we get the tail. Assign new head and tail.
Since we start from the head, formula used for 'for' loop is n - (k % n)
- 1. This approach does not change the value of k. It starts from the
head. It uses for loop to iterate instead of while so has smaller code.