Floyd cycle detection algorithm
Iterate over the entire first linked list. We can make the first linked
list into a cycle by attaching the last element to the first element. We
are making a ring. Assign the fast and slow pointers to the second list
and find the place of entry of the cycle. Linked List Cycle 2 shows the
explanation. The place where the cycle enters is where intersection
begins and rest of the elements are common for both lists. Before
returning the entry point i.e intersection, make next element after last
element of first linked list as None so that we do not modify the
structure of list.