Two pointers approach
The benefit of this approach is because of saving space. We traverse
both linked lists in until both lists intersect. The lists are traversed
two times if both do not intersect as both will have null value.
Time complexity is O(2(N+M)) i.e O(N+M) and space complexity is O(1).
The condition to mention if pointer is None pass it to next list else
pass it to next element of same list. We can return either pointer of
both lists.