DFS with recursion
Time complexity: O(N) traversing all the nodes in the tree.
Space: O(N) for the space used by solution array

Here we use dfs with recursion to traverse all the nodes.
We define an output solution array and use it to store the nodes based
on their height. We call recursively the dfs method with root as input.
Return the output array at the end. The DFS method does following -
Leaf nodes get height of 0(-1 + 1) because they have no children so
their left and right children return -1.  If the height of node is
greater or equal to length of output array, then append an empty list []
to the output. Place the root value at the proper location in output.
Return the height of root node.