DFS iterative solution
Time and space O(n) same as recursive. Here we use stack instead of
recursive dfs method inside given method. Root and max so far become the
stack array's elements. Since base case of null node cannot be written,
therefore we check is not a null nodes using if statement for node and
its children. Usual while loop for stack is run and current node is
popped. Add the left and right child to the stack along with max_so_far.
Return the no of good nodes.