Elementary math approach using dummy and carry
Since the input is already reversed i.e lowest digit first, we need to
just add the digits in both list. To take care of the scenario wherein
digit does not exist in the list for that position, we write node.val if
node else 0.

We create a new ListNode to store the output. Before the head of the
output, we add a dummy node. Return head.next at end.

To add both lists, we use the while condition where atleast one element
exists in either of the lists.
Edge cases to be considered:
1. One of the lists becomes empty, l1 = [], l2 = [1,2,3]
2. Sum is greater than 10, so carry = 1, l1 = [1], l2 = [9,9,9]
3. Length of lists are different.

Use different pointers for both lists l1 and l2 (p1 and p2), and for
output (head and dummy). Keep moving dummy and p1, p2 after the addition
is done and new node is added. Add carry as ListNode if more than 0.