Using neetcode solution of brute force Along with dfs and constrained programming i.e writing conditions to place a dot.
Instead of 3, 4 dots considered to know when to stop and return or when
to add string to result.

Three arguments needed while calling recursively the helper function,
index, dots and current IP string. To start recursive function, we pass
arguments as (index = 0, dots = 0, curIP = ''). For next element, pass
(index + 1, dots + 1, curIP + string from i to j and 1 dot) in recursive
function.

While appending to result, we do not add a copy of current IP address
because strings do not support copy method and we can not modify an
existing string. So we create new string by passing it as an argument in
the recursive function (curIP + s[i: j + 1] + '.').

Since for loop runs from i to i + 3, we consider 3 possibilities, i, i+1
and i + 2, maybe that is why popping/ Backtracking is not needed as we
append to result only when 4 dots (dots == 4) are present and index has
reached end of string (i == len(s)).

Constraints for next recursive call inside for loop to place dot are:
1. string from i to j converted to int is less than 256 (s[i:j+1]) < 256
2. Integer is either 0 (i == j) or first number is not 0 (s[i] != '0')

Condition (if dots > 4: return) stops unnecessary iterations such as
4.5.6.1.25.1.13 early.

if len(s) < 3 or len(s) > 12: return, does not consider IP addresses of
incorrect length.