300 Longest Increasing Subsequence
8 8 2022

Dynamic Programming and Greedy with binary search approaches

time: O(N log N) using binary search
space: O(N) for lis array, utlimately return len of this array

DP and Greedy both take O(N^2) time and O(N) space.
N^2 because we iterate through each element and decide whether to use it
or not. If we decide to use it then we need to iterate over the entire
output array to check where to place it. This second linear search for
location can be improved using binary search.