109 Convert Sorted List to Binary Search Tree
8 11 2022

Inorder traversal in iterative way
We pass start, end, status alongwith root node in stack as arguments.
Status is used because we are doing inorder traversal, we will pass
root node again to stack but with flag status as True. When the flag is
True, assign the value of linked list's current node and move pointer of
linked list to the next node. Add left and right nodes to the stack only
when the values of start <= mid - 1 for left and mid + 1 <= end. This is
with right side elements only so for [-10,-3,0,5,9] linked list, bst
created will be [0,-10,5,null,-3,null,9] with traversal as following:
0 4 False
0 1 False
0 1 True
1 1 False
1 1 True
0 4 True
3 4 False
3 4 True
4 4 False
4 4 True

Time is O(N) as we visit each node in linked list atleast once and space
for stack is O(N)